==> Python Scripts/01_data_cleaning.ipynb <==
# Data Cleaning

# importing libraries and loading data

import pandas as pd
import numpy as np
import os
import janitor

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/data'
products = pd.read_csv(os.path.join(path, 'products.csv'), index_col=False)
orders = pd.read_csv(os.path.join(path, 'orders.csv'), index_col=False)
departments = pd.read_csv(os.path.join(path, 'departments.csv'))

# wrangling departments csv

departments = (departments
               .T
               #.reset_index()
)
new_header = departments.iloc[0]
departments = departments[1:]
departments.columns = new_header

# creating data dict
data_dict = departments.to_dict('index')
 ##### not sure this is right please let me know! 

# wrangling orders csv

orders = orders.drop(columns=['eval_set'])

orders = orders.change_type(['order_id','user_id'], str)

## Further Data Cleaning

# updating col names
(orders
    .rename_column(old_column_name = 'order_dow', new_column_name= 'order_day_of_week')
    .rename_column(old_column_name = 'days_since_prior_order', new_column_name= 'days_since_last_order')
)

# it looks like the busiest hour is 10am
orders.value_counts('order_hour_of_day')

print(data_dict.get('4'))

# creating a subset of breakfast item sales

breakfast = products[products['department_id'] == 14]


# creating a subset of possible dinner parties

parties = products.loc[products['department_id'].isin([5,7,12,20])]

# this df has  7688 entries 

# a df about user 1

user_1 = orders[orders['user_id'] == '1']

user_1.describe()

# extracting clean data 
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
orders.to_csv(os.path.join(path, 'orders_wrangled.csv'), index=False )
departments.to_csv(os.path.join(path, 'departments_wrangled.csv'), index=False)
