==> Python Scripts/02_data_cleaning.ipynb <==
# Cleaning up Orders Data

# loading modules
import pandas as pd
import numpy as np
import os 
import janitor

# loading data
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'

orders = pd.read_csv(os.path.join(path, 'orders_wrangled.csv'))

orders.describe()

1. it looks like we are missing some enteries in the 'days since prior order' column, since the count is lower than the others

2. There doesn't seem to be any other issues otherwise 

orders[orders['order_number'] > 21]

# looking for mixed type data

for col in orders.columns.tolist():
  weird = (orders[[col]].applymap(type) != orders[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (orders[weird]) > 0:
    print (col)

# it looks like there are no mixed-type data

# looking for missing values

orders.isnull().sum()

orders[orders['days_since_prior_order'].isnull() == True]

it looks like the dataframe only keeps track of one record per user_id of when the last day since prior order was. 

to this end it is okay to leave the NaN's as is and just be sure to exclude them when making calculations specific to this column. 

# looking for duplicates

orders[orders.duplicated()]

it looks like there are no duplicates and so we have nothing to deal with here!

# exporting data
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
orders.to_csv(os.path.join(path, "orders_wrangled.csv"), index=False)
