==> Python Scripts/03_merging_tables.ipynb <==
# Joining Tables

# setting up modules

import pandas as pd
import numpy as np
import os
import janitor

# loading in data

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/data'
orders_prior = pd.read_csv(os.path.join(path,'order_products__prior.csv'))
products = pd.read_csv(os.path.join(path,"products.csv"))


## data checks for products table

## checking for nan's and removing them
products.isnull().sum()

clean_products = products[products['product_id'].isnull() == False]

## looking for dups
clean_products[clean_products.duplicated()]

cleaner_products = clean_products.drop_duplicates()

## seeing if product id is still duplicated 

cleaner_products[cleaner_products.product_id.duplicated()]

# exporting to csv
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
cleaner_products.to_csv(os.path.join(path, 'products_wrangled.csv'))

## data checks for order tables

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
orders = pd.read_csv(os.path.join(path,'orders_wrangled.csv'))

orders_prior.shape
orders_prior.describe()

# seeing if there are any duplicated in order id
orders.order_id.duplicated().sum()

orders.describe()

## merging and exporting data

# merging the two dataframes

orders_large = orders.merge(orders_prior, on = 'order_id', indicator = True)

orders_large._merge.value_counts()

# exporting data
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
orders_large.to_pickle(os.path.join(path, 'orders_products_combined.pkl'))

# reading in data
orders_full = pd.read_pickle(os.path.join(path,'orders_products_combined.pkl')).drop(columns = ['_merge'])

products = pd.read_csv(os.path.join(path, 'products_wrangled.csv'))

# there are two products with the same product id 32404859

products[products.product_id.isin([6799,6800,26519,26520,26521])]

#pd.merge(orders_full,products, how='left', on = 'product_id', indicator = True).value_counts('_merge')

#pd.merge(products,orders_full, how='left', on = 'product_id', indicator = True).value_counts('_merge')

orders_products_merged = orders_full.merge(products, how='inner', on='product_id')

# exporting the data
orders_products_merged.to_pickle(os.path.join(path, 'orders_products_merged.pkl'))
