==> Python Scripts/04_creating_analysis_var.ipynb <==
# Creating New Variables in orders_products_merged

# importing libraries

import pandas as pd
import numpy as np
import os
import janitor

# loading in data

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
opm = pd.read_pickle(os.path.join(path, 'orders_products_merged.pkl'))
df = opm[:1000000]

#opm = opm.drop(columns='Unnamed: 0')

opm.head()

### creating price range descriptions

def price_label(row):

    if row['prices'] <= 5:
        return 'low-range-product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'mid-range product' 
    elif row['prices'] > 15:
        return 'high range' 
    else: return 'not enough data'

df['price_range'] = df.apply(price_label, axis = 1) # axis 1 = rows

df.price_range.value_counts()

# doing the same thing but using loc on the full data set
opm.loc[opm['prices'] > 15, 'price_range_loc'] = 'high-range product'

opm.loc[(opm['prices'] <= 15) & (opm['prices'] > 5), 'price_range_loc'] = 'mid-range product'

opm.loc[opm['prices'] <= 5, 'price_range_loc'] = 'low-range product'

### creating for loop to display busiest day

result = []

for value in opm['order_dow']:
    if value == 0:
        result.append('busiest day')
    elif value == 4:
        result.append('least busy day')
    else: 
        result.append('regular busy day')

#opm['busiest_day'] = result
#opm['busiest_day'].value_counts(dropna = False)

## updating busiest day col

opm.order_dow.value_counts()

opm = opm.case_when(
    ((opm.order_dow == 0) | (opm.order_dow == 1)), 'busiest days',
    ((opm.order_dow == 3) | (opm.order_dow == 4)), 'slowest days',
    'regular days',
    column_name = 'busiest_days'
)


# cheacking the values for this new col for accuracy
opm.busiest_days.value_counts()

## creating col to show busiest time periods of the day

list_of_top_hours = opm.order_hour_of_day.value_counts().index.tolist()

opm.loc[opm['order_hour_of_day'].isin(list_of_top_hours[0:8]), 'busiest_period_of_day'] = 'most orders'
opm.loc[opm['order_hour_of_day'].isin(list_of_top_hours[8:16]), 'busiest_period_of_day'] = 'average orders'
opm.loc[opm['order_hour_of_day'].isin(list_of_top_hours[16:24]), 'busiest_period_of_day'] = 'fewest orders'
opm.loc[~opm['order_hour_of_day'].isin(list_of_top_hours), 'busiest_period_of_day'] = 'not enough data'

opm.busiest_period_of_day.value_counts()

# exporting df

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
opm.to_pickle(os.path.join(path, 'opm.pkl'))
