==> Python Scripts/05_creating_analysis_var.ipynb <==
# Creating Analysis Variables for OPM dataset

# importing modules
import pandas as pd
import numpy as np
import os
import janitor

# loading in data
path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
opm = pd.read_pickle(os.path.join(path,'opm.pkl'))

#  cleaning extreme price values
opm.loc[opm['prices'] > 100, 'prices'] = np.nan

# finding mean order_number by department groups
(opm
.groupby('department_id')
.agg({'order_number':['mean']})
)

compared to the result in the excercise, with this result we now can see all of the department id's. In addition, on average, the averages are higher than those produced with the subset. 

I'm not entirely sure this measure is informative? Given that the order_number just states which order number sequentially it is for a specific user. I believe we are not gaining any measure of quantity, but rather a measure of sequentiality.

opm['max_order'] = (opm
                    .groupby(['user_id'])['order_number']
                    .transform(np.max))

opm.loc[opm['max_order'] > 40, 'loyalty_flag'] = 'loyal customer'
opm.loc[(opm['max_order'] <= 40) & (opm['max_order'] > 10), 'loyalty_flag'] = 'regular customer'
opm.loc[opm['max_order'] <= 10, 'loyalty_flag'] = 'new customer'

opm[['user_id','max_order','loyalty_flag']].tail(10)

opm.tail(20)

# finding 
(opm
.groupby('loyalty_flag')
.agg({'prices':['mean','max','min']})
)

# creating a new col with average prices
opm['average_price'] = (opm
                    .groupby(['user_id'])['prices']
                    .transform(np.mean))

# checking to see if calculation was done correctly
opm[opm['user_id']==203436]['prices'].mean()

opm.loc[opm['user_id'] == 203436,['user_id','prices','average_price']]

# creating a new col to flag customers

opm.loc[opm['average_price'] < 10, 'spender_type'] = 'low spender'
opm.loc[opm['average_price'] >= 10, 'spender_type'] = 'high spender'

# creating median days since ordered col per user

opm['median_day_since_ordered'] = (opm
                    .groupby(['user_id'])['days_since_prior_order']
                    .transform(np.median))

# checking to see if calc were done correctly
opm[opm['user_id']==202557]['days_since_prior_order'].median()

# creating a new col to flag frequent customers
opm.loc[opm['median_day_since_ordered'] <= 10, 'customer_freq'] = 'frequent customer'
opm.loc[(opm['median_day_since_ordered'] > 10) & (opm['median_day_since_ordered'] <=20), 'customer_freq'] = 'regular customer'
opm.loc[opm['median_day_since_ordered'] > 20, 'customer_freq'] = 'non-frequent customer'

# exporting data
opm.to_pickle(os.path.join(path,'opm.pkl'))
