==> Python Scripts/06_data_prep_for_viz.ipynb <==
# Preping Data for Visualizing Instacart Data

import pandas as pd
import numpy as np
import os
import janitor
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/data'
customers = pd.read_csv(os.path.join(path,'customers.csv'))

## data checks for customer table

customers.head()

# changing col names 
customers = customers.clean_names()

customers.info()

# checking for duplicates
customers.duplicated().sum()

# checking for nan's 
customers[customers['first_name'].isnull()]

# we are missing some first names in the data, but that is not much of a worry
# and we cannot do anything about it

# checking for mixed data types
for col in customers.columns.tolist():
  weird = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# first_name is the only col that pops up most likely because of the NaN's

## merging dataframes

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
opm = pd.read_pickle(os.path.join(path, 'opm.pkl'))

opm.describe()

help = opm.merge(customers, on= 'user_id', indicator= True)

help.to_pickle(os.path.join(path,'final.pkl'))

## testing to see if file exported properly

path = r'/Users/johntam/cf_data_analysis/tasks/A4/code/clean_data'
test = pd.read_pickle(os.path.join(path,'final.pkl'))

final = test.drop(columns = ['_merge'], axis = 1)

final.head()
